+++ packages/amplify-graphql-auth-transformer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.12](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-auth-transformer@4.1.11...@aws-amplify/graphql-auth-transformer@4.1.12) (2025-02-07)
+
+**Note:** Version bump only for package @aws-amplify/graphql-auth-transformer
+
 ## [4.1.11](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-auth-transformer@4.1.10...@aws-amplify/graphql-auth-transformer@4.1.11) (2025-01-30)
 
 ### Bug Fixes