+++ packages/amplify-graphql-schema-generator/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.7](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.11.6...@aws-amplify/graphql-schema-generator@0.11.7) (2025-02-07)
+
+**Note:** Version bump only for package @aws-amplify/graphql-schema-generator
+
 ## [0.11.6](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.11.5...@aws-amplify/graphql-schema-generator@0.11.6) (2025-01-30)
 
 **Note:** Version bump only for package @aws-amplify/graphql-schema-generator